+++ eva/src/components/dashBoardInside/dashConstructorScheme/dashConstructorSchemes.vue
@@ -19,10 +19,12 @@
         class="dash-constructor-schemes__options"
         :class="{
           'dash-constructor-schemes__options--is-keymap-open': isKeymapOpen,
+          'dash-constructor-schemes__options--edit-mode': isEdit,
+          'dash-constructor-schemes__options--dnd-panel-is-open': dndPanel,
         }"
       >
         <v-tooltip
-          v-if="dashboardEditMode"
+          v-if="false"
           bottom
           :color="theme.$accent_ui_color"
         >
@@ -32,6 +34,7 @@
                 v-model="isEdit"
                 inset
                 dense
+                class="dash-constructor-schemes__switch"
                 label=""
                 @change="toggleInputMode"
               />
@@ -39,25 +42,84 @@
           </template>
           <span>Вкл\выкл режим редактирования</span>
         </v-tooltip>
-        <template v-if="isEdit && dashboardEditMode">
+        <template v-if="dashboardEditMode">
           <v-tooltip
             bottom
             :color="theme.$accent_ui_color"
           >
             <template v-slot:activator="{ on }">
-              <div class="pa-2 d-flex">
+              <div class="pa-5 d-flex">
                 <v-icon
                   class="control-button edit-icon theme--dark"
                   :style="{ color: theme.$secondary_text }"
                   v-on="on"
                   @click="toggleDnDPanel"
                 >
-                  {{ gear }}
+                  {{ dndPanel ? arrowCollapseLeft : arrowCollapseRight }}
                 </v-icon>
               </div>
             </template>
             <span>Панель настроек</span>
           </v-tooltip>
+          <template v-if="searchForBuildScheme">
+            <v-tooltip
+              :disabled="isLoading"
+              bottom
+              :color="theme.$accent_ui_color"
+            >
+              <template v-slot:activator="{ on }">
+                <div class="pa-2 d-flex">
+                  <v-icon
+                    class="control-button edit-icon theme--dark"
+                    :style="{ color: theme.$secondary_text }"
+                    v-on="on"
+                    @click="openConfirmModal"
+                  >
+                    {{ cloudDownloadOutlineIcon }}
+                  </v-icon>
+                </div>
+              </template>
+              <span>Загрузить с сервера</span>
+            </v-tooltip>
+          </template>
+          <template v-if="saveMultipleScheme">
+            <v-select
+              v-model="localActiveSchemeId"
+              :items="allSavedSchemes"
+              class="
+                dash-constructor-schemes__select-field
+                dash-constructor-schemes__select-field--with-padding
+              "
+              label="Активная схема"
+              dense
+              :menu-props="{
+                'offset-y': true,
+              }"
+            >
+              <template #item="{ item, on }">
+                <v-list-item
+                  ripple
+                  class="v-list-item--link"
+                  v-on="on"
+                >
+                  <v-list-item-content>
+                    <v-list-item-title class="d-flex align-center justify-content-between">
+                      <div class="mr-auto">
+                        {{ item }}
+                      </div>
+                      <v-icon
+                        class="control-button edit-icon theme--dark"
+                        :style="{ color: theme.$secondary_text }"
+                        @click.stop="openConfirmDeleteModal(item)"
+                      >
+                        {{ closeIcon }}
+                      </v-icon>
+                    </v-list-item-title>
+                  </v-list-item-content>
+                </v-list-item>
+              </template>
+            </v-select>
+          </template>
           <template v-if="dataSelectedNode">
             <v-tooltip
               bottom
@@ -124,27 +186,6 @@
               <span>На уровень ниже</span>
             </v-tooltip>
           </template>
-          <template v-if="searchForBuildScheme">
-            <v-tooltip
-              :disabled="isLoading"
-              bottom
-              :color="theme.$accent_ui_color"
-            >
-              <template v-slot:activator="{ on }">
-                <div class="pa-2 d-flex">
-                  <v-icon
-                    class="control-button edit-icon theme--dark"
-                    :style="{ color: theme.$secondary_text }"
-                    v-on="on"
-                    @click="openConfirmModal"
-                  >
-                    {{ cloudDownloadOutlineIcon }}
-                  </v-icon>
-                </div>
-              </template>
-              <span>Загрузить с сервера</span>
-            </v-tooltip>
-          </template>
         </template>
         <v-tooltip
           bottom
@@ -205,14 +246,16 @@
         }"
         :class="{
           'dash-constructor-schemes__dnd-panel-container--active': dndPanel,
-        // 'dash-constructor-schemes__dnd-panel-container--is-keymap-open': isKeymapOpen
         }"
       >
         <div
           v-show="!isLoading"
+          v-if="false"
           class="row justify-end"
         >
-          <div class="col-auto">
+          <div
+            class="col-auto"
+          >
             <button
               @click="toggleDnDPanel"
             >
@@ -560,6 +603,14 @@
         btn-cancel-text="Нет"
         @result="startSearch"
       />
+      <modal-confirm
+        v-model="isConfirmModalDelete"
+        :theme="theme"
+        :modal-text="`Удалить выбранную схему из списка сохраненных ?`"
+        btn-confirm-text="Да"
+        btn-cancel-text="Нет"
+        @result="deleteSavedScheme"
+      />
     </div>
   </portal>
 </template>
@@ -569,6 +620,8 @@ import {
   mdiArrowDown,
   mdiArrowUp,
   mdiClose,
+  mdiArrowCollapseLeft,
+  mdiArrowCollapseRight,
   mdiSettings,
   mdiHelp,
   mdiFitToPageOutline,
@@ -638,6 +691,8 @@ export default {
       gear: mdiSettings,
       closeIcon: mdiClose,
       arrowUp: mdiArrowUp,
+      arrowCollapseLeft: mdiArrowCollapseLeft,
+      arrowCollapseRight: mdiArrowCollapseRight,
       iconArrowUp: '/icons/OrderIcons/bring_to_front.svg',
       arrowDown: mdiArrowDown,
       iconHelp: mdiHelp,
@@ -705,6 +760,9 @@ export default {
       // activeScheme: 'graph',
       timeout: null,
       timer: 0,
+      isConfirmModalDelete: false,
+      schemeIdForDelete: '',
+      localActiveSchemeId: '',
     };
   },
   computed: {
@@ -732,7 +790,7 @@ export default {
     savedGraphObject() {
       const savedGraph = this.dashFromStore[this.idFrom]?.savedGraphObject;
       if (savedGraph) {
-        return savedGraph[this.activeSchemeId] || [];
+        return savedGraph[this.localActiveSchemeId] || [];
       }
       return [];
     },
@@ -790,12 +848,15 @@ export default {
             result.push(tokenByName.value);
           }
         });
-        return result.join('-').replaceAll(' ', '_') || 'graph';
+        return result.join('-').replaceAll(' ', '_') || 'default-scheme';
       }
-      return 'graph';
+      return 'default-scheme';
     },
     allSavedSchemes() {
-      return Object.keys(this.dashFromStore[this.idFrom].savedGraphObject);
+      if (this.dashFromStore[this.idFrom]?.savedGraphObject) {
+        return Object.keys(this.dashFromStore[this.idFrom].savedGraphObject);
+      }
+      return [];
     },
   },
   watch: {
@@ -824,9 +885,8 @@ export default {
       });
     },
     dashboardEditMode(val) {
-      if (!val) {
-        this.isEdit = false;
-      }
+      this.isEdit = val;
+      this.toggleInputMode();
     },
     dataForBuildScheme(value) {
       if (this.isAlwaysUpdateScheme && value?.length > 0) {
@@ -837,7 +897,13 @@ export default {
         );
       }
     },
-    activeSchemeId() {
+    activeSchemeId(schemeId) {
+      this.localActiveSchemeId = schemeId;
+      if (!this.dashFromStore[this.idFrom].savedGraphObject[schemeId]) {
+        this.updateSavedGraphObject([]);
+      }
+    },
+    localActiveSchemeId() {
       if (!this.isAlwaysUpdateScheme && this.constructorSchemes) {
         this.constructorSchemes.update(this.savedGraphObject);
       }
@@ -846,11 +912,19 @@ export default {
       this.toggleLoading(val);
     },
   },
+  created() {
+    if (!this.dashFromStore[this.idFrom].savedGraphObject) {
+      this.localActiveSchemeId = this.activeSchemeId || 'default-scheme';
+      this.createSavedGraphObjectField();
+    }
+  },
   mounted() {
     this.createGraph();
     this.updateDefaultElementColor = throttle(this.updateDefaultElementColor, 200);
     this.updateSavedGraph = throttle(this.updateSavedGraph, 1000);
     this.setActions();
+
+    this.isEdit = this.dashboardEditMode;
   },
   methods: {
     setActions() {
@@ -910,7 +984,7 @@ export default {
         updateStoreCallback: this.updateSavedGraph,
         updateStoreCallbackV2: this.updateSavedGraphObject,
         toggleLoadingCallback: this.toggleLoading,
-        isEdit: this.isEdit,
+        isEdit: this.dashboardEditMode,
         isBridgesEnable: this.isBridgeEnable,
         onClickObject: (type, data) => {
           if (type !== 'label-type-0') {
@@ -987,10 +1061,10 @@ export default {
           value: {},
         }]);
       }
-      if (!this.dashFromStore[this.idFrom]?.savedGraphObject[this.activeSchemeId]) {
+      if (!this.dashFromStore[this.idFrom]?.savedGraphObject[this.localActiveSchemeId]) {
         this.$store.commit('setState', [{
           object: this.dashFromStore[this.idFrom].savedGraphObject,
-          prop: this.activeSchemeId,
+          prop: this.localActiveSchemeId || 'default-scheme',
           value: [],
         }]);
       }
@@ -1004,7 +1078,7 @@ export default {
         this.createSavedGraphObjectField();
         this.$store.commit('setState', [{
           object: this.dashFromStore[this.idFrom].savedGraphObject,
-          prop: this.activeSchemeId,
+          prop: this.localActiveSchemeId,
           value: data,
         }]);
         if (this.dashFromStore[this.idFrom].savedGraph || this.dashFromStore.savedGraph) {
@@ -1056,6 +1130,7 @@ export default {
     toggleInputMode() {
       if (this.constructorSchemes) {
         this.isEdit = this.constructorSchemes.toggleInputMode();
+        // this.localActiveSchemeId = this.activeSchemeId;
         if (!this.isEdit) {
           this.closeDataPanel();
           this.dndPanel = false;
@@ -1092,22 +1167,39 @@ export default {
       if (confirm) {
         this.constructorSchemes.buildSchemeFromSearch(this.dataForBuildScheme);
       }
-      // if (confirm) {
-      //   this.isConfirmUpdateScheme = true;
-      //   this.$store.commit('updateSearchStatus', {
-      //     idDash: this.idDashFrom,
-      //     id: this.searchForBuildScheme,
-      //     status: 'empty',
-      //   });
-      // } else {
-      //   this.isConfirmModal = false;
-      // }
     },
     updateIconsList(iconsListFrom) {
       return iconsListFrom
         .filter((elementFrom) => !this.primitivesFromStore
           .some((element) => element.icon === elementFrom.icon))
         .map((element) => element.icon);
+    },
+    openConfirmDeleteModal(schemeIdForDelete) {
+      this.schemeIdForDelete = schemeIdForDelete;
+      this.isConfirmModalDelete = true;
+    },
+    deleteSavedScheme(isCancelDelete) {
+      if (isCancelDelete) {
+        const deleteFn = (object, fieldsForDelete) => {
+          const result = {};
+          Object.keys(object).forEach((key) => {
+            if (!fieldsForDelete.includes(key)) {
+              result[key] = object[key];
+            }
+          });
+          return result;
+        };
+        const allSchemes = structuredClone(this.dashFromStore[this.idFrom].savedGraphObject);
+        const filteredSavedSchemes = deleteFn(allSchemes, [this.schemeIdForDelete]);
+        this.$store.commit('setState', [{
+          object: this.dashFromStore[this.idFrom],
+          prop: 'savedGraphObject',
+          value: filteredSavedSchemes,
+        }]);
+        this.schemeIdForDelete = '';
+      } else {
+        this.schemeIdForDelete = '';
+      }
     },
   },
 };
@@ -1133,8 +1225,28 @@ export default {
     }
   }
   &__select-field {
+    &--with-padding {
+      padding-left: 10px;
+      padding-right: 10px;
+    }
     ::v-deep {
-      border-color: var(--main_text);
+      &.v-input__control, .v-icon__svg {
+        caret-color: var(--main_text) !important;
+        color: var(--main_text) !important;
+        border-color: var(--main_text) !important;
+      }
+    }
+  }
+  &__switch {
+    padding-left: 20px;
+    ::v-deep.v-input__control {
+      .v-input__slot {
+        .v-input--selection-controls__input {
+          .v-input--switch__track {
+            color: var(--secondary_text);
+          }
+        }
+      }
     }
   }
   &__loading-circular {
@@ -1146,11 +1258,32 @@ export default {
   }
   &__options {
     position: absolute;
-    left: 20px;
-    top: 0;
+    left: 0;
+    top: 5px;
     display: flex;
     align-items: center;
     z-index: 10;
+    padding-right: 5px;
+    transition: all .2s ease;
+    &--edit-mode {
+      &::before {
+        content: "";
+        position: absolute;
+        pointer-events: none;
+        left: 0;
+        top: 0;
+        right: 0;
+        bottom: 0;
+        transition: all .2s ease;
+        background-color: var(--main_bg);
+        opacity: .8;
+        z-index: -1;
+      }
+    }
+
+    &--dnd-panel-is-open {
+      left: 255px;
+    }
   }
   &__inner-options {
     display: flex;